separate mobile and desktop styles
